Update patch set 4
Patch Set 4: Code-Review+1

whitespace has been cleaned,

regarding PhoneCapabilityTester, that was a mistake on my part as it is a privately contructable class in Contacts app and hence cannot be used here. 

Also since the user is already on the Dialpad, we can be sure that TELEPHONY_SERVICE and FEATURE_TELEPHONY are present, which itself makes sure that SMS service is present. (The only way to make sure of SMS capability is checking TELEPHONY_SERVICE and FEATURE_TELEPHONY only.)

There can be devices with SMS and not dialling ability, but there aren't any devices with dialling, but no SMS facility.

Patch-set: 4
Reviewer: Gerrit User 1024916 <1024916@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1