Update patch set 1
Patch Set 1:

Has the internal tree already addressed this issue ?

If not, this patch (or something similar) is totally necessary. Currently you cannot tap the "All contacts" button in most situations, mainly because there is a reference to that view stored while listview is trying to recycle it.

Whats even worse is, any input events get queued and dispatched when you try to leave the app.

Patch-set: 1