Update patch set 1
Patch Set 1:

p.pawit: i don't get a resource not found error building AOSP, so i don't think this fixes anything. (but i agree with nancychen it fixes a style issue.)

nancychen: note that with AOSP only googlers see a "Submit" button, so you have to +1 Verified the change and hit Submit too.



but... i do see a crash:

  04-26 21:57:41.134  4967  4967 E AndroidRuntime: Process: com.android.dialer, PID: 4967
  04-26 21:57:41.134  4967  4967 E AndroidRuntime: java.lang.NullPointerException: Attempt to invoke virtual method 'java.lang.String java.lang.Object.toString()' on a null object reference
  04-26 21:57:41.134  4967  4967 E AndroidRuntime: 	at android.widget.ArrayAdapter.createViewFromResource(ArrayAdapter.java:394)
  ...

and a little debugging shows it's not because of this; it's because you're passing String[] { null }. and that's because -- presumably because i have no SIM? -- you're being told that there's a phone, but that its device id is null.

Patch-set: 1