Update patch set 1
Patch Set 1:

crash is not good, but I don't see enough of the crash log to know where the problem is. Elliott, could you post more of the stack trace or the fix? re: localization, it would be better to guard against the NPE with a null check and even if localization is not needed to put the string in strings.xml resources file

Patch-set: 1