Update patch set 1
Patch Set 1:

I'm building AOSP to test it.

From readin the code, it looks like we have an "ok" string in our non-AOSP builds but not our AOSP builds so it does seem like this is a bug.  Will dig into this and if it fails, which I suspect it will, we're likely to take this bug.

Patch-set: 1
Reviewer: Gerrit User 1015328 <1015328@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0