Update patch set 1
Patch Set 1: Code-Review-2

Thank you for catching this. The dialer team has reviewed and taken the steps to submit this change that will be merged into AOSP in the near future. As this was an important fix, the dialer team decided to take a direction intended to ensure that the fix is taken up by all internal branches.

Patch-set: 1