Automatically respond to upgrade request from VT-TX to bidirectional
When an upgrade request from one-way video (VT-TX) to bidirectional
video is received, a video response screen is displayed. Displaying
the video response screen is not needed if the call is already in
video state.

So if the previous call state is video, automatically respond to
upgrade request.

Test: manual - Checked the video call screen after receiving the
upgrade request from one-way video (VT-TX) to bidirectional video.
Bug: 63605284

Change-Id: I2c652cd404ecc1ba160ce2fdbd959c3981b57b38