Show audio button with checked color when Bluetooth connected
This solves the usability issue that user can't recognize easily
if Bluetooth device is connected during a call.
When the Bluetooth device is connected and call audio route is
Bluetooth or speaker, set the checked state to the audio button
and show it with checked color.

Test: manual - Checked that the Bluetooth button and speaker button are
shown with checked color when Bluetooth connected
Bug: TBD

Change-Id: I60c45b1a419cd722a329491b92df18e5fd0876f6