Update patch set 2
Patch Set 2:

Hi Kyle, does the issue being addressed in this CL present itself after every rotation or changing of the camera or is it less easy to reproduce? If it's not quite so easy, could you briefly describe how we could produce the incorrect behavior? A bug link would also be helpful one exists. Thanks!

Patch-set: 2
CC: Gerrit User 1168623 <1168623@85c56323-6fa9-3386-8a01-6480fb634889>