Update patch set 2
Patch Set 2:

Sorry for the slow reply Kyle.

The AOSP Dialer code here is a couple of years out of date and is obsolete (Dialer has not released updated source code in a couple of years). We will be releasing updated Dialer source code with Android O.

I have confirmed the bugs addressed here are fixed in Dialer so they will also be fixed in the upcoming Dialer open source update.

Patch-set: 2