Fixed issue where Voicemail could not be played
The root cause is that the termination process of
VoicemailPlaybackPresenter is called and mContext is null.
Because VoicemailPlaybackPresenter is singleton.

When Dialer app is lauched from Notification,
the existed DialtactsActivity is destroyed
and new DialtactsActivity is created.

When the existed DialtactsActivity is destroyed,
onDestroy of VoicemailPlaybackPresenter is called.

However, when new DialtactsActivity is created,
VoicemailPlaybackPresenter is not initialized.
So, this issue is reproduced.

Make counter measures to prevent unnecessary clear processing by
making reference counters in VoicemailPlaybackPresenter.

Affected area:
 Voicemail display
 Voicemail playback

Bug: 69324209
Change-Id: Ie9dbf7580417a9ff1cea84bf9468a7f0758d57f3