Fix showing post digits on In-Call UI
The object "mPrimary" is always the same outgoing call for single
MO call so that the object reference isn't changed, which means the
flag "primaryChanged" is always true, so that the primary call info
can't be updated after lower layer updates the address.

To fix it, add an extra condition for "shouldRefreshPrimaryInfo" that
checks whether the address is equal with the number of mPrimaryContactInfo.

Change-Id: I3cd0722bc86ebba4537106bbfaef350cf350d5ab