Fix issue voice screen is shown for a short time when making video call
It can't judge whether a call is video call or not by
DialerCall#isVideoCall before a connection is created. So, a voice call
screen is shown by mistake.
Use VideoProfile#isVideo instead of DialerCall#isVideoCall to judge a
call state properly.

Test: manual - Checked a voice call screen is not shown after making a
video call and a video call screen is shown immediately.
Bug: 67720489

Change-Id: Id0bd3261bdd78fe7920c2141932ccedf21e94ddc