IMS-VT: Fix crash for incoming VT call on top of upgrade request.
Preview surface during incoming VT call should be handled by
AnswerFragment and the corresponding VideoCallPresenter.
However VideoCallPresenter associated with VideoCallFragment is
wrongly handling the incoming call event just after it was
torn down. This causes local preview surface to be attached to
textureview of VideoCallFragment and results in a crash due to
wrong state of surface texture.
To fix this, ignore call state events in VideoCallPresenter when
UI is not ready.

Test: Manual

Bug: 63608380
Change-Id: I7a5f3d50a70a4fff4d6fd0c45d7fa4207379d290