Update patch set 1
Patch Set 1:

Hi and thanks for the CL. Using SystemClock instead is reasonable. My concern is that the values we get from https://developer.android.com/reference/android/telecom/Call.Details.html#getConnectTimeMillis() are not based on elapsed time. If this is the case, the duration value calculated will not be the expected value. I'll look into this and comment with my findings.

Patch-set: 1
CC: Gerrit User 1168623 <1168623@85c56323-6fa9-3386-8a01-6480fb634889>