Fix wrong orientation change when a video call ends
When video call is disconnected,
VideoCallPresenter#checkForOrientationAllowedChange is invoked with
"Call(= null)". When Call is null, set "DISALLOW" to
Activity#setAllowOrientationChange. It causes that the end call screen
is displayed in portrait mode even if in landscape mode.

When Call is null, do not invoke
VideoCallPresenter#checkForOrientationAllowedChange. It avoid to be set
"DISALLOW" improperly.

Test: manual - Checked that orientation is not rotated when video call
ends.
Bug: 80375554

Change-Id: I2f1d474429b2b5abbc95179352d2205aebbf95c6