Improved behavior of back button in search ui.
Pressing the back button in the search UI now functions as follows:
 - If the keyboard is opened, the keyboard is minimized
 - If the dialpad is opened, the dialpad is closed
 - If the keyboard and dialpad is closed, the search ui is closed

Our existing behavior was dependent on whether a query had been built yet.
basically, if the user pressed back with no query selected, the search ui
was closed.

From the bugbash:
7. No scroll bar in the search results if the results do not fill the entire
view. The keyboard overlaps the results and there is no way to get to the
enter list without dismissing the keyboard.
10. Dismiss the keyboard on tapping the down arrow on the keyboard. Arrow
points down but works like back button

Bug: 64902476,64137632,62685859,63691995,63939331
Test: manual
PiperOrigin-RevId: 167332236
Change-Id: I1c0b5b429316dae119b5fb21be4303d7fe052e35