Automated g4 rollback of changelist 167332236.
*** Reason for rollback ***

Based off a CL that needs to be rolled back.

*** Original change description ***

Improved behavior of back button in search ui.

Pressing the back button in the search UI now functions as follows:
 - If the keyboard is opened, the keyboard is minimized
 - If the dialpad is opened, the dialpad is closed
 - If the keyboard and dialpad is closed, the search ui is closed

Our existing behavior was dependent on whether a query had been built yet.
basically, if the user pressed back with no query selected, the search ui
was closed.

From the bugbash:
7. No scroll bar in the search...

***

Bug: 64902476,64137632,62685859,63691995,63939331
Test: n/a
PiperOrigin-RevId: 167661409
Change-Id: If657c47f5496925273e4564251cc4a80015017de