Remove voicemail count changed check for legacy notifications
Previously before showing a voicemail notification dialer checks if the count has changed to avoid showing duplicated notifications. With the dismissed flag in place the count changed check is redundant, and it prevents the notification to reshow after reboot.

Also updated the EXTRA_IS_REFRESH to use the O_MR public version.

Bug: 66633426
Test: LegacyVoicemailNotificationReceiverTest
PiperOrigin-RevId: 171760393
Change-Id: I508b8f42a44fd849ab7c8ca6fe9d4e7e8a38b1c7