Fix issue user can't recognize "Hanging up" label when a call is ending
"Hanging up" is displayed with the fade-in animation during
disconnecting call and then, "Call ended" is displayed with the
animation when the call is disconnected. The duration of disconnecting
status is very short and the call status is changed to disconnected.
User can't recognize "Hanging up" label because "Call ended" is
displayed in the middle of fade-in animation of "Hanging up" label, 
Especially, this is bad user experience if "Hanging up" is translated in
so long.

To solve this problem, stop the animation if a call status is
disconnecting or disconnected.

Test: manual - Checked that the call state label is correctly shown
when Ims call is ending
Bug: TBD

Change-Id: Ic2676d8f0fdfbb6caf831863b1e757c08211b063