Update patch set 1
Patch Set 1:

There is no interface from RIL/Modem to telephony to call this api.
One option is to have this implementation in TelephonyConnection.java in packages/services/Telephony. keep a Boolean to track whether answer request came or not, and if without receiving answer request, call state active comes then call the telecom manager api to bring incallui to foreground.
Is this approach OK?

Patch-set: 1