Update patch set 1
Patch Set 1:

(1 comment)

Tyler, could we use android.provider.CallLog.Calls.FEATURES_HD_CALL to determine if HD icon should be shown in call logs?

Is it currently actually being used?
IIRC there is issue with that this bit is set even it's not HD voice call. I'm not sure if this is also affected in call logs.

Patch-set: 1