Add expand and collapse media player logic for NUI Voicemail Fragment
This adds a file called new_voicemail_media_player_layout.xml that can be expanded and collapsed on tapping the voicemail entry. This file will later be updated in accordance with the specs, but the logic for expansion and collapsing would not be affected by the future cosmetic change.

Video=https://drive.google.com/open?id=1aIpmVfUYrFrg_FZ4tRX_TX6pU9YgPQaN
Bug: 64882313,68382421
Test: Unit Tests
PiperOrigin-RevId: 174713873
Change-Id: I828719541d42b5d3162701be9229ffae5a5900b3