Guard against remote directories returning results with no phone numbers during search.
We have observed in the wild that some apps implementing remote directory don't seem to respect the where clause.

Note that if a search returns 10 results, and 4 of them have null numbers, we will only show 6 numbers where ideally we'd show 10. (It would take a lot of work to fix that just to workaround misbehaving providers which I don't think is worthwhile.)

Bug: 68942044
Test: unit + manually changed guard condition and observed rows hidden
PiperOrigin-RevId: 176183008
Change-Id: Icdb26fa043644e27a4b6ac3c767c728b3ca224ff