Fix showing CallingAccountSelector in lock screen
android:showWhenLocked and FLAG_SHOW_WHEN_LOCKED is added to PreCallActivity so it can be launched in lockscreen.
PreCallCoordinator now grabs new instances of PreCallActions onResume so the state can be flushed. If it is reused states like PendingAction might be used after it is discarded.

Added null check before discarding fragment because it might not be created yet due to background query.

Bug: 69570769
Test: CallingAccountSelectorTest
PiperOrigin-RevId: 176587256
Change-Id: Iaa7145e860136e9177f568e2c84461ed5efa12a1