Contact list no longer jumps to the top when the cp2 database updates.
The contacts fragment recycler view was rebuilding the adapter and layout
manager each time the Contacts.CONTENT_URI observered a content change. Now we
simply update the adapter's cursor and notifyDataSetChanged. This keeps the
recycler view in place through all changes (inserts, deletes and updates).

This change also fixes another bug replated to contacts fragment not refreshing
when the contacts permission was granted in another part of the UI.

Bug: 64044576,67781478,68179085
Test: manual
PiperOrigin-RevId: 177106761
Change-Id: Ib6a929efb047001d681cb008c3ede69860146836