Delete voicemails when disabling visual voicemail
For privacy reasons we now delete all voicemails and transcriptions when
the user disables visual voicemail, (after confirming that they want to do this
via a dialog).

Note: we only delete voicemails whose source package is the google dialer.
Note: the voicemails will be re-downloaded and re-transcribed if the user re-enables visual voicemail

Bug: 69323147
Test: manual and unit test
PiperOrigin-RevId: 177315139
Change-Id: Ie04496dc5960b485794733fbf74f7f704e806023