Add Assisted Dialing Call Details Implementation.
This includes all of the work specified in the UX layouts,
with the exception of the parenthetical country code.

Because the canonical mapping of dial prefixes to country
codes is not contained in this package, some thought needs to
be given on the best way to implement this part of the spec.

This change also introduces a fallback error text, in case the
client experiences bad data from the call log.

Test: new unit tests
PiperOrigin-RevId: 178038123
Change-Id: I88eec72a73820e092f24c5f53ee9520a42486ada