Register content observer when voicemail table changes.
When a new voicemail is received, it is written in the voicemail table by the Voicemail service. However the new voicemail will not get updated/shown in the New Voicemail UI as the annotated call log would be stale. This CL ensures that when the voicemail is added, the annotated call log is marked dirty and refreshed. This way the new voicemail will be shown in the new voicemail UI. Since a new voicemail is also added, we want to make sure the headers for "today" and "older", their positions are also updated accordingly.

Bug: 64882313
Test: Unit tests
PiperOrigin-RevId: 179623267
Change-Id: I5b70cac91a9219dbb8ee722fd76f22ccef39718f