Added copySubMessage method to PhoneLookup interface.
The existing way that protos are merged in CompositePhoneLookup is not correct because foo_submessage from BarDataSource may incorrectly contribute old information to the merged message.

The new copySubMessage method makes it so that each PhoneLookup is responsible for defining which submessage it is responsible for and prevents the problem.

Test: unit
PiperOrigin-RevId: 179707015
Change-Id: I9a31a573968ba0709d73835d17b335c8a6501a68