Add assisted dialing indicators to the old call log.
Modify the call log grouping strategy. The strategy is as follows:

When comparing the current group to the next candidate to add to that group,
if the assisted dialing feature tag differs between the two, create a new
group.

Thus, if a call has other features like wifi or hd, we ignore those features
when making grouping decisions.

Bug: 70506228
Test: unit tests
PiperOrigin-RevId: 180592125
Change-Id: I73d130bd9eb23706a04cb02f5711200729d978b6