Added Search Bar to NUI.
This change only adds the views but doesn't implement any logic
or functionality. That will come in a future CL. A decision was
made here about whether to create a new search bar or reuse the
old one. I opted to make a new one to avoid breaking the old one
while it's still being used.

Bug: 64655802
Test: n/a
PiperOrigin-RevId: 181512198
Change-Id: Ib377b1d99f4eb8100204ac9a110dc5899bced392