Write PhoneLookup results to PhoneLookupHistory in RealtimeRowProcessor.
This is an optimization to reduce "popping" in the new call log. Since we have to do the PhoneLookup anyway to update the UI, we may as well write the result to PhoneLookupHistory so that the next time the AnnotatedCallLog is refreshed, the updated results can be retrieved from PhoneLookupHistory.

I also updated RealtimeRowProcessorTest to use FakePhoneLookup rather than the real Cp2PhoneLookup since RealtimeRowProcessor no longer uses Cp2PhoneLookup directly (it was updated to use the general-purpose PhoneLookup in a previous CL but I didn't update the test at that time).

Test: unit
PiperOrigin-RevId: 182974567
Change-Id: I813e9d69f802ca08757238290fdfcf58e78b3592