Fixed a regression where we incorrectly reported that the dialpad was visible.
A request was made to update the value of #isDialpadVisible, in the process a
regression was introduced that would cause #isDialpadVisible to return true
when the dialpad's root view was null. This caused a NPE when we tried to
animate it.

We should now return the correct value in the scenarios important to our
partners as well our users and animations.

Bug: 64657832,72384544
Test: manual/logging
PiperOrigin-RevId: 183122628
Change-Id: I981e944dbb003064163b8ec11883dbc5381cba0e