Implemented multiselect in NUI for VisualVoicemailFragment.
This solution has a couple hacks in it because the DialtactsActivity has a
non-support ActionBar and MainActivity has a support Toolbar. When multiselect
is implemented for NewVoicemailFragment, we should be careful to use support
action mode instead of non-support action mode so that the transitions and
styling are seamless and consistent.

Bug: 72722083
Test: manual
PiperOrigin-RevId: 184343414
Change-Id: Id16652de2d0add7b7c304d1cedb429fe9b1cf338