Update patch set 18
Patch Set 18: Presubmit-Verified+1

=== Finished presubmit run: L91500000142984637 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/611866?revision=18&workplanId=L91500000142984637




TreeHugger finished with: 15 passed.







WARNING: The tip-of-tree(4596661) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4596497), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.








WARNING: The tip-of-tree(4596661) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4596446), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4596661) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4596497), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







Congratulations!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 18
Tag: autogenerated:TreeHugger:Verify-Pass
Label: Presubmit-Verified=+1