Use UI Listeners for querying voicemail status table
Earlier we were using the legacy CallLogQueryHandler for querying the voicemail status table. However we should be using UI Listeners instead in NUI. This CL queries the voicemail status table and returns a list of voicemail status', which are then used to display the corresponding error messages.

This CL also moved VoicemailStatus out of the legacy dialer/database and moves it to dialer/voicemailstatus and ensures that there are no dependencies on the legacy dialer/database in the NUI.

Bug: 73139237
Test: Unit tests
PiperOrigin-RevId: 185321023
Change-Id: Id16ea475b6a52da380fbf8b3590dc75cbcdc370e