Mark calls as read in new call log.
Playing with the existing app, the missed call becomes unbolded when:

1) Expanding the row. The closest analog of this is in the new UI is opening the bottom sheet, I've done that.

2) Swiping away from the call history tab. This can't be done in NewCallLogFragment because it doesn't know if the user is selected a new tab or pressed Home. So, I implemented this in NewMainActivityPeer.

3) After viewing the call log for 3(ish) seconds and leaving the activity pressing Home/Back. This is best done in NewCallLogFragment#onResume since MainActivity doesn't always know when the fragment is being displayed (it could be done after the user comes back to the app after pressing Home for example).

Note that the notification is also removed in all of these cases.

Also note that dismissing the notification makes the call unbolded (but this case already appears to be handled via the system call log content observer).

Also, as part of writing tests for this, I made TestCallLogProvider more realistic.

Bug: 70989622
Test: manual
PiperOrigin-RevId: 185457438
Change-Id: Ib360d3bc73083bd1a018ed98e2b7d9a69fb7fafb