Add "enabled" to SelectPhoneAccountDialogFragment.
This is used to inform the user an account cannot be used right now. On most dual SIM devices, only a single SIM can make calls at the same time. The UI will be implemented in a followup CL.

This CL also packs the parameters of SelectPhoneAccountDialogFragment into a proto. There are too many arguments and it needs structured representation.

TEST=TAP
Bug: 69675796,72618783
Test: TAP
PiperOrigin-RevId: 194139636
Change-Id: I7d9f92c73b650654fff28ba625a2c8e3dfa0b96c