Update patch set 2
Patch Set 2:

>Good suggestion. I will investigate if we are able to solve the problem using BidiTextView.

I have now tried to refactor the solution to use BidiTextView instead. Unfortunately it did not work flawlessly out of the box. When using BidiTextView a space is missing between the phone number and the custom label. I have not uploaded the refactored solution due to this problem.

We will continue to investigate this problem.

Patch-set: 2
Reviewer: Gerrit User 1233564 <1233564@85c56323-6fa9-3386-8a01-6480fb634889>