Fixed speed dial remove.
App would crash when users attempted to remove a starred contact because we try
to update Phone.CONTENT_URI, which is not allowed. Test assumed that
we could update PHone.CONTENT_URI, which is why this wasn't caught.

Updated the test to verify the contact in Contacts.CONTENT_URI get updated to
starred = 0. Unforunately, since fakeContentProvider doesn't work exactly like
cp2, the update won't propogate to Phone.CONTENT_URI, so out test is less than
ideal, but it should cover our bases.

Also updated the image and text for starring a suggestion.

Bug: 78923466
Test: tap
PiperOrigin-RevId: 194962080
Change-Id: I9f8f71d70ba83510682890917081380a3c51cd7a