Set the DisplayNameSource to PHONE in DefaultLookupUriGenerator.
In this case we don't have a name, but set the DISPLAY_NAME to the number to make the contact card look nice. However, when using STRUCTURED_NAME and creating a new contact with the number, the number is populated in both the first/last name fields and the number field. Using PHONE makes it so the number is displayed on the contact card but the number is not populated in the name fields when creating a contact. This is consistent with the old call log's behavior (see ContactInfoHelper:createTemporaryContactUri).

TEST=manual
Bug: 78252466
Test: manual
PiperOrigin-RevId: 195714742
Change-Id: I26cf18a76af60f48d7979cd8cc8e47d72de1bf5e