Check if SIM for voicemail account is present
When a SIM with visual voicemail account has been removed
while phone turned off, the voicemail tab is still present
in Dialer and voicemail messages are visible. If SIM
removed while phone turned on, the voicemail account is
removed and this issue is not seen.

Add a check if SIM for voicemail account is still present
when counting number of voicemail accounts. Number of
voicemail accounts are used to judge whether voicemail
tab should be present in Dialer UI.

Bug: 80204634
Test: Manual
Change-Id: Ic76165d2c4ebe72f38298e335c02e22727e8670b