Register system call log content observer if user enables Phone permission.
Previously, we only attempted to create the content observer in Application#onCreate and if the user didn't have the permission enabled, it wouldn't get added and would remain off for the duration of the application's life. Now we check if the observer is registered when refreshing the call log and enable it if necessary.

TEST=unit
Bug: 72461366
Test: unit
PiperOrigin-RevId: 196543060
Change-Id: I1d58efd21fb63e0745b43ac8ff11d87562126a2f