Fix add call button behavior.
Double clicking add call button mutes the call, and going back to in call UI doesn't unmute the call. It's because previousMuteState is set twice. We shouldn't do anything for the second click since the first click makes you leave in call UI.

Also change automaticallyMuted to automaticallyMutedByAddCall to clearly indicate the value is for add call button only.

Test: CallButtonPresenterTest
PiperOrigin-RevId: 196590319
Change-Id: I9f41b1a75ced3900ae0c4fa787f3defaf7f1cbe6