Cleaned up deps in SpamBlockingPromoHelperTest.
Deprecated RobolectricApplication.

RobolectricApplication uses RobolectricDialerRootComponent, which depends on all parts of the dialer application (CallLog, PhoneLookup, CommandLine, etc). None of these things matter for leafy unit tests like SpamBlockingPromoHelperTest and they just serve to slow down builds; both for tests that use RobolectricApplication and presubmits for all parts of the application. (Unrelated changes to the call log for example, would trigger this test on presubmit because this test depends on call log when it shouldn't.)

TEST=yes
Test: yes
PiperOrigin-RevId: 197193370
Change-Id: I425351eef2954d105ac7fc7c4120767d275a4fc8