Update patch set 2
Patch Set 2:

This change only hides the tab, we need a more robust solution that can actually disable VVM (which includes cleanup like deleting all local VMs.) This will be tricky to do because the SIM is often not "present" when the phone is booting up, and we don't want to redownload every voicemail on reboot. We do plan to work on this but there are no set schedule yet.

Patch-set: 2
Reviewer: Gerrit User 1097644 <1097644@85c56323-6fa9-3386-8a01-6480fb634889>