Update patch set 3
Patch Set 3:

Hi, we have now found the cause of the problem with the missing space.

The string is segmented in DialerBidiFormatter.java. For example the string "<string1> +1 650-253-0000 <string2>" will be segmented into ["<string1> ","+1 650-253-0000"," <string2>"]. If <string1> is written in an RTL-language the whitespace after <string1> will be placed first in the resulting string. The resulting string will therefore be " <string1>+1 650-253-0000 <string2>".

Our solution is to also segment the whitespaces that surrounds the segments above by performing modifications in DialerBidiFormatter.java. The string "<string1> +1 650-253-0000 <string2>" would therefore be segmented into ["<string1>"," ","+1 650-253-0000"," ","<string2>"] This forces the whitespace to be placed correctly in relation to the phone number, and the resulting string will be "<string1> +1 650-253-0000 <string2>

Patch-set: 3
Reviewer: Gerrit User 1233564 <1233564@85c56323-6fa9-3386-8a01-6480fb634889>