Fix unnatural animation while move up "Swipe down to reject"
The animation of "Swipe down to reject" moves unnaturally from third
times at receiving screen. It moves upward suddenly when the third
bounce animation starts. This behavior occurs for "Swipe down to reject"
does not have animation which is moving up in bounce animation. So,
the start position of its animation is not initialized.

This patch adds the animation which is moving up to
"Swipe down to reject".

Test: manual - Check that "Swipe down to reject" moves like the
animation of the first time and second time, it looks naturally.
Bug: TBD

Change-Id: Id8eafa2daf1ebb5de10eb3bca42c5d311862558b