Fix call provider label not update when redial emergency call
When redialing an emergency call by other sub, the call provider label
is displayed with the old sub info. The call provider label is cached,
however the cache is not updated in the emergency call redialing.
Since the DialerCall object has not been destroyed across redialing,
the cache of old sub will be used as it is.

To resolve this issue, clear cache of call provider label when
PhoneAccountHande is updated. By clearing the cache, gets the correct
info from PhoneAccount for the next time.

Bug: 113303907
Test: manual - Checked that call provider label is changed correctly
when redialing an emergency call
Change-Id: Ic17523f8d32fb990dccc6748772c923ab47b169b