Delete voicemail account for removed SIM
When a SIM with visual voicemail account has been removed
while phone turned off, the voicemail tab is still present
in Dialer since voicemail account is not removed.

Add a check if SIM for voicemail account is still present.
If SIM is not present, delete voicemail account and
voicemails belonging to that account.

Bug: 80204634
Test: Manual
Change-Id: Ic76165d2c4ebe72f38298e335c02e22727e8670b