Update patch set 3
Patch Set 3:

We have now uploaded a new patchset that we think handles previous review comment.

We have changed the functionality of this patch to not only hide the voicemail tab. Instead, vvm account and voicemails are deleted if the corresponding SIM card is no longer present.

Is this what you had in mind?

Patch-set: 3
Reviewer: Gerrit User 1233564 <1233564@85c56323-6fa9-3386-8a01-6480fb634889>