WIP - [GMS Dialer] listen to proximity sensor state and filter touch
This is just a quick hack to see if it is possible to filter touch
events based on proximity sensor state.  This does not fully fix
the problem but is intended for discussion purposes.

Without this fix it is possible to mute, end call and other such
things with your cheek/ear etc.

Bug: 116076184
Test: Using a Pixel device make a call, move the phone away from
      head upwards (imagine leaning forwards or turning the device away
      to talk to someone). Put the phone back to ear. Call should not
      mute.
      An easier way to reproduce this is to:
      1. Make a call
      2. Touch and hold the mute button
      3. Cover the proximity sensor and wait for 1 second
      The InCallUI should not receive the mute button ACTION_UP
      when the screen is off and the finger is still on the screen.
Change-Id: If5ab7143fedcd80b3ed8edec07f6b7cbed831fed