Disable resource deduping
AAPT2 resource deduping removes mcc mnc
specific values when those are the same as the default value.
This causes problems when the language is something else.

This Bug was found when language is set to German and you
use a Vodafone DE sim card with MCC262 MNC02.
Vodafone wants the name for notification_call_wifi_brand to be
Wifi Call but you get WLAN-Anruf since the mcc262mnc2 resource
is removed in the deduping.

Test: make Dialer and check resources with "aapt2 dump resources"
Bug: 171892595
Change-Id: I791f93e3226002d89c2a756105014647750dd093